geo_credit_score_ai/pyproject.toml
[project]
name = "fraud-detection-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imbalanced-learn>=0.14.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.0",
]

[tool.ruff]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg",
    ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d",
    ".pyenv", ".pytest_cache", ".pytype", ".ruff_cache", ".svn",
    ".tox", ".venv", ".vscode", "__pypackages__", "_build",
    "buck-out", "build", "dist", "node_modules", "site-packages", "venv",
]

line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E", "F", "I", "B", "C4", "UP", "W",
]
ignore = [
    "E501",    # line length
    "B905",    # zip() without strict
    "TRY003",  # overly verbose exceptions
    "ANN",     # type annotations not enforced
]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "ignore::UserWarning:sklearn.utils.validation",
]
